Refactor: UI 컴포넌트와 상태 관리 로직(useMemoModal) 분리
